refactor: simplify payments component names
