--- frontend/src/features/payments/index.ts
@@ -1,4 +1,0 @@
-export * from '@/features/payments/stores/payment.store'
-export * from '@/features/payments/stores/paymentInteraction.store'
-export * from '@/features/payments/stores/paymentInfo.store'
-export * from '@/features/payments/types'
