+++ frontend/src/payments/components/CurrencySelectorDialog.vue
