+++ frontend/src/payments/stores/paymentInteraction.store.ts
@@ -2,12 +2,12 @@ import { computed, ref } from 'vue'
 import { defineStore, storeToRefs } from 'pinia'
 
 import { useI18nStore } from '@/localization/store'
-import { waitForDeepLinkLaunch } from '@/features/payments/services/deepLinkService'
-import { usePaymentInfoStore } from '@/features/payments/stores/paymentInfo.store'
-import { usePaymentStore } from '@/features/payments/stores/payment.store'
-import type { DeepLinkProvider, PaymentMethod } from '@/features/payments/types'
-import { createPaymentActionResolver } from '@/features/payments/workflows/createPaymentActionResolver'
-import type { PaymentActionContext } from '@/features/payments/workflows/types'
+import { waitForDeepLinkLaunch } from '@/payments/services/deepLinkService'
+import { usePaymentInfoStore } from '@/payments/stores/paymentInfo.store'
+import { usePaymentStore } from '@/payments/stores/payment.store'
+import type { DeepLinkProvider, PaymentMethod } from '@/payments/types'
+import { createPaymentActionResolver } from '@/payments/workflows/createPaymentActionResolver'
+import type { PaymentActionContext } from '@/payments/workflows/types'
 import { isMobileDevice } from '@/shared/utils/device'
 import { openUrlInNewTab } from '@/shared/utils/navigation'
