+++ frontend/src/payments/workflows/actions/default.ts
@@ -1,4 +1,4 @@
-import type { PaymentActionContext, PaymentMethodAction } from '@/features/payments/workflows/types'
+import type { PaymentActionContext, PaymentMethodAction } from '@/payments/workflows/types'
 
 export const createDefaultAction = (context: PaymentActionContext): PaymentMethodAction => ({
   handleSelection: ({ method, currency }) => {
